Add files via upload
In this project, we will focus on salary prediction. The data set for this assignment includes information on job descriptions and salaries. Use this data set to see if we can predict the salary of a job posting (i.e., the `Salary` column in the data set) based on the job description. This is important because this model can make a salary recommendation as soon as a job description is entered into a system.